More tests and debug output.
Correct 2 clear cases of wrong 64-bit tests and another one that
apparently would work either way.

Revert one case of considering a register dead after an non local
unconditional branch, because the code could be reached with a
local branch, and in that case, lightning will know the register
is live.